requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
django-cors-headers==4.3.1
Pillow==10.1.0

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-celery-beat==2.5.0

# WebSockets / ASGI
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Servers
gunicorn==21.2.0
uvicorn[standard]==0.24.0

# Auth / Security
PyJWT==2.8.0
cryptography==41.0.7
django-filter

# HTTP
requests==2.31.0
httpx==0.25.2
yfinance

# Data Processing (Python 3.12 compatible)
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1

# Machine Learning (Python 3.12 compatible)
tensorflow==2.16.1
keras==3.3.3
scikit-learn==1.4.2
statsmodels==0.14.1

# Technical Analysis
ta==0.11.0
# TA-Lib requires manual installation of system library
# Remove the below line unless you have TA-Lib installed on your OS
# ta-lib==0.4.28

# NLP
textblob==0.17.1
nltk==3.8.1

# LLM / AI Integrations
google-generativeai==0.3.2
langchain==0.1.20
langchain-google-genai==0.0.5

# Streaming / Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Spark
pyspark==3.5.1
# py4j removed (already included in pyspark)

# HDFS
hdfs==2.7.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0


# # Testing
# pytest==7.4.3
# pytest-django==4.7.0
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# factory-boy==3.3.0
# faker==20.1.0

# # Development
# black==23.12.0
# flake8==6.1.0
# isort==5.13.2
# pre-commit==3.6.0
# django-debug-toolbar==4.2.0
# django-extensions==3.2.3
